Update de Students

+++ src/database/migrations/20191030165531-create-students.js
@@ -25,7 +25,7 @@ module.exports = {
                 allowNull: false,
             },
             stature: {
-                type: Sequelize.INTEGER,
+                type: Sequelize.FLOAT(),
                 allowNull: false,
             },
             created_at: {
